Broke out ManageButton for memberships into its own file
